Defend against missing Annotation/Reply body graph nodes and blobs.
git-svn-id: http://svn.topazproject.org/svn/head@7205 48088309-b40f-0410-b149-b14953253675

+++ ambra/webapp/src/main/java/org/topazproject/ambra/annotation/service/AnnotationService.java
@@ -160,7 +160,7 @@ public class AnnotationService extends BaseAnnotationService {
     pep.checkAccess(AnnotationsPEP.DELETE_ANNOTATION, URI.create(annotationId));
     ArticleAnnotation a = session.get(ArticleAnnotation.class, annotationId);
 
-    if (a != null) {
+    if ((a != null) && (a.getBody() != null)) {
       String pp[]     = new String[] { a.getBody().getId() };
       session.delete(a);
       permissionsService.cancelPropagatePermissions(annotationId, pp);
